Decompose DataColumnSidecarCustodyImpl: extract long polling functionality to a separate DataColumnSidecarCustody implementation
